adding logic to repriotize new account

+++ grails-app/services/net/hedtech/banner/general/overall/DirectDepositAccountCompositeService.groovy
@@ -452,26 +452,36 @@ class DirectDepositAccountCompositeService {
 
     def rePrioritizeAccounts( def map, def newPosition ) {
 
-        def accountList = directDepositAccountService.getActiveHrAccounts(map.pidm)
-        accountList.sort {it.priority};
-
         def reOrderInd = true
         def priorityList = []
 
         def adjustedMapList = []
         def prioritizedList = []
 
-        //make sure new position is not greater than the number of hr accounts.
-     //   def newPositionExceedsInd = false
-        if (newPosition > accountList.size()) {
-            reOrderInd = false
-        }
+        def newAcct = false
 
-        //if record with "remaining" (param is map) is sent for re prioritizing
-        //then no action will be taken. No records will be updated.
         def itemBeingAdjusted = map
-        if (itemBeingAdjusted.percent == 100) {
-            newPosition = accountList.size()
+
+        //checking for mandatory values
+        //  if ()
+
+        def accountList = directDepositAccountService.getActiveHrAccounts(itemBeingAdjusted?.pidm)
+        accountList.sort {it.priority};
+
+
+        if(!itemBeingAdjusted.containsKey("id") || itemBeingAdjusted?.id == null) {
+            itemBeingAdjusted.priority = setNextPriority(itemBeingAdjusted).priority
+            itemBeingAdjusted.id = -1
+            newAcct = true
+
+            def domainObject = new DirectDepositAccount()
+            def routingObject = new BankRoutingInfo()
+            domainObject.properties = itemBeingAdjusted
+            routingObject.properties = itemBeingAdjusted.bankRoutingInfo
+            domainObject.bankRoutingInfo = routingObject
+            domainObject.id = -1
+
+            accountList << domainObject
         }
 
         //if the new position that is sent, happens to be the position of "Remaining" record, then
@@ -479,21 +489,42 @@ class DirectDepositAccountCompositeService {
         //for example, if new position and the position of "remaining" record is 6, then the
         //new position will be 5.
         //should not adjust the position of "remaining" record.
-        def remainingPosition = accountList.findIndexOf  { iterator ->
+        def remainingPosition = (accountList.findIndexOf  { iterator ->
             iterator.percent == 100
+        })+1
+        //produce an error, if the item being adjusted is at 100% and also remaining exists already
+        if (remainingPosition > 0) {
+            if (newPosition == remainingPosition || newPosition > remainingPosition) {
+                newPosition=newPosition-1;
+            }
         }
-        remainingPosition = remainingPosition+1
 
-        if (newPosition == remainingPosition) {
-            newPosition=newPosition-1;
+        //if record with "remaining" (param is map) is sent for re prioritizing
+        //then no action will be taken. No records will be updated.
+        if (itemBeingAdjusted?.percent == 100) {
+            newPosition = accountList.size()
+        }
+
+        //make sure new position is not greater than the number of hr accounts.
+        if (!newAcct) {
+            if (newPosition > accountList.size()) {
+                reOrderInd = false
+            }
         }
 
         //if the new position is same as the position of the item being adjusted,
         //then no reordering will be done.
-        def itemBeingAdjPosition = accountList.findIndexOf  { iterator ->
-            iterator.id == map.id
+        def positionBeingUpdated
+        if (newAcct) {
+            positionBeingUpdated=accountList.size()
+
+        } else {
+            positionBeingUpdated = (accountList.findIndexOf  { iterator ->
+                iterator.id == itemBeingAdjusted?.id
+            })+1
         }
-        if (newPosition == itemBeingAdjPosition+1) {
+
+        if (newPosition == positionBeingUpdated) {
             reOrderInd = false
         }
 
@@ -501,25 +532,17 @@ class DirectDepositAccountCompositeService {
         //if the original record set has a record with "remaining" or 100%
         //then, eliminate that record from the list.
         //record with remaining should not be re prioritized.
-        accountList.remove(accountList.find { p -> p.percent == 100 })
+//        accountList.remove(accountList.find { p -> p.percent == 100 })
         accountList.sort {it.priority};
 
-        //finding the position of the record being updated from input param
-        //adding  +1, as index starts from 0.
-        def positionBeingUpdated = accountList.findIndexOf  { iterator ->
-            iterator.id == map.id
-        }
-        positionBeingUpdated = positionBeingUpdated+1
 
-        //if the input param is "remaining" no updates or no reordering will be done.
         if (reOrderInd) {
             //assign the new position to the item being adjusted and add it to the list.
             def adjItem = [:]
             adjItem.id = itemBeingAdjusted.id
             adjItem.newPosition = newPosition
             adjustedMapList << adjItem
-            def origPriority = (accountList.find { p -> p.id == itemBeingAdjusted.id } as DirectDepositAccount).priority
-            priorityList << origPriority
+            priorityList << (accountList.find { p -> p.id == itemBeingAdjusted.id } as DirectDepositAccount).priority
 
             if (newPosition > positionBeingUpdated) {
                 for (int i=newPosition; i>positionBeingUpdated;  i--) {
@@ -555,34 +578,37 @@ class DirectDepositAccountCompositeService {
         adjustedMapList.each {
             def pid=it.id
             def acct = accountList.find { p -> p.id == pid } as DirectDepositAccount
-            directDepositAccountService.delete(acct)
+            if (pid != -1) {
+                directDepositAccountService.delete(acct)
+            }
+
         }
 
         int i = 0
         adjustedMapList.each {
             def pid=it.id
-            def acct = accountList.find { p -> p.id == pid } as DirectDepositAccount
-            if (acct.id == map.id) {
-                acct.accountType = map.accountType
+            def acct = accountList.find { p -> p.id == pid }
+            if (acct.id == itemBeingAdjusted.id) {
+                acct.accountType = itemBeingAdjusted.accountType
 
-                if (map.percent != null && map.percent != "") {
-                    if (map.percent.getClass() == String) {
-                        acct.percent = Double.parseDouble(map.percent)
+                if (itemBeingAdjusted.percent != null && itemBeingAdjusted.percent != "") {
+                    if (itemBeingAdjusted.percent.getClass() == String) {
+                        acct.percent = Double.parseDouble(itemBeingAdjusted.percent)
                     } else {
-                        acct.percent = map.percent
+                        acct.percent = itemBeingAdjusted.percent
                     }
                     acct.amount = null
                 }
 
-                if (map.amount != null && map.amount != "") {
-                    if (map.amount.getClass() == String) {
-                        acct.amount = Double.parseDouble(map.amount)
+                if (itemBeingAdjusted.amount != null && itemBeingAdjusted.amount != "") {
+                    if (itemBeingAdjusted.amount.getClass() == String) {
+                        acct.amount = Double.parseDouble(itemBeingAdjusted.amount)
                     } else {
-                        acct.amount = map.amount
+                        acct.amount = itemBeingAdjusted.amount
                     }
                     acct.percent = null
                 }
-             }
+            }
             acct.priority = priorityList[i++]
             prioritizedList << acct
 
@@ -599,5 +625,6 @@ class DirectDepositAccountCompositeService {
 
         return directDepositAccountService.getActiveHrAccounts(map.pidm)
     }
+
 
 }
